+++ Main.py
@@ -21,15 +21,16 @@ import Utilities._check_input as CheckInput # Verify the argparsed elements
 # IMPORTATION OF ALL THE SYSTEM OF EQUATIONS 
 #!#!#! FOR THE MOMENT ONLY ONE SYSTEM IS IMPORTED 
 import SystemOfEquations.GK_Noinflation as GK_Noinflation
-import SystemOfEquations.G_Reduced as G_Reduced
+
 
 # Default parameter values for main control
 _DEFAULTPARAMETERS = {
-            'SystemOfEquation'  : G_Reduced.MODEL(),
+            'SystemOfEquation'  : GK_Noinflation.MODEL(),
             'InitialDictionnary': None,
             'Value_Changes'     :{'NumParameters'      : {},
                                   'Initial conditions' : {},
                                   'Parameters'         : {},
+                                  'ParNum'             : {},
                                   'TypeOfArrays'       : 'lin' #log
                                   },
             'plot'              : True,
@@ -38,7 +39,7 @@ _DEFAULTPARAMETERS = {
             'savepath'          : os.path.dirname(__file__),
             'returnAll'         : False,
             'Comments'          : "",
-                     }
+                    }
 
 
 # #############################################################################
@@ -130,7 +131,7 @@ def run(
         tim = time.time()
         print('Start simulation...', end='')
 
-    y = CORE.initializeY(initCond) # vector y contains all states at time t.
+    y = CORE.initializeY(initCond) # vector y contains all states of time t.
     Y_s, t_s = CORE.TemporalLoop(y) # Calculation of all timesteps
 
     if timeit is True:
