+++ NotIncludedYet/_def_parameters.py
@@ -1,34 +1,36 @@
 # -*- coding: utf-8 -*-
 """
-PARAMETER POOL FOR SIMULATIONS
+PARAMETER POOL FOR SIMULATIONS 
 
-All the parameters are stocked in _DPARAM.
+All the parameters are stocked in _DPARAM. 
 None is the default value if the key has no meaning for the parameter
     Each parameter is a dictionary, with the following keys :
     'NAME': {                           # KEY THE CODE WILL USE TO CALL IT
         'value': 100,                   # NUMERICAL VALUE
-        'com': 'Duration of simulation',# Commentary about what it means
-        'dimension': 'time',            # Physical dimension if relevant
+        'name': 'time steps',           
+        'com': 'Duration of simulation',# Commentary about what it means 
+        'dimension': 'time',            # Physical dimension if relevant 
         'units': 'y',                   # Physical units if relevant
-        'type': None,                   # Intensive or extensive
-        'symbol': None,                 # Plot-friendly name (Latex)
+        'type': None,                   # Intensive or extensive 
+        'symbol': None,                 # Plot-friendly name. If None, it'll be 'NAME''
         'group': 'Numerical',           # For practical regroupment of variables
-    },
+    },       
 
 ###############################
 
-PARAMETERS :
+PARAMETERS : 
     _PARAMSET    : Name of the set of default parameter set taken by the system. 
+    _FLATTEN     : !!! UNUSED FOR THE MOMENT
     _DALLOWED    : List of types and dimensions accepted by the system (with None)
     _DEF_PARAM   : All the informations about useful parameters
     _dfail       : Parameters that couldn't get loaded because incomplete
     _lkeys       : List of attributes necessary for a parameter to be added
     _DPARAM      : Presets of parameters !!! NOT CODED FOR THE MOMENT
 
-FUNCTIONS:
+FUNCTIONS :    
     _check_inputs: Check if the the input of the user is in term of format
     get_params   : Description linked to the function
-
+    
 """
 
 
@@ -36,6 +38,7 @@ import numpy as np
 
 
 _PARAMSET = 'v0'
+_FLATTEN = True
 
 
 # ---------
@@ -62,7 +65,6 @@ _PARAMSET = 'v0'
 # ic['L'] = ic['lambda']*ic['N']
 # ic['W'] = ic['omega']*ic['a']
 
-
 _DALLOWED = {
     'dimension': ['time', 'time rate', 'temperature rate'],
     'type': ['intensive', 'extensive'],
@@ -70,15 +72,14 @@ _DALLOWED = {
 
 
 # ##########################################
-# PARAMETERS
-
-
+# PARAMETERS PARAMETERS
 _DEF_PARAM = {
 
     # --------------
     # Numerical
     'Tmax': {
         'value': 100,
+        'name': 'time steps',
         'com': 'Duration of simulation',
         'dimension': 'time',
         'units': 'y',
@@ -88,6 +89,7 @@ _DEF_PARAM = {
     },
     'Nx': {
         'value': 1,
+        'name': None,
         'com': 'Number of similar systems evolving in parrallel',
         'dimension': None,
         'units': None,
@@ -97,6 +99,7 @@ _DEF_PARAM = {
     },
     'dt': {
         'value': 0.01,
+        'name': None,
         'com': 'Time step (fixed timestep method)',
         'dimension': 'time',
         'units': 't',
@@ -106,6 +109,7 @@ _DEF_PARAM = {
     },
     'Tstore': {
         'value': None,  # Dynamically allocated
+        'name': None,
         'com': 'Time between storages (if StorageMode=full, it goes to dt)',
         'dimension': 'time',
         'units': None,
@@ -115,6 +119,7 @@ _DEF_PARAM = {
     },
     'Nt': {
         'value': None,  # Dynamically allocated
+        'name': None,
         'com': 'Number of temporal iteration',
         'dimension': None,
         'units': None,
@@ -124,6 +129,7 @@ _DEF_PARAM = {
     },
     'Ns': {
         'value': None,  # Dynamically allocated
+        'name': None,
         'com': 'Number of elements stored',
         'dimension': None,
         'units': None,
@@ -133,6 +139,7 @@ _DEF_PARAM = {
     },
     'verb': {
         'value': True,
+        'name': None,
         'com': 'flag indicating whether to print intermediate info',
         'dimension': None,
         'units': None,
@@ -142,6 +149,7 @@ _DEF_PARAM = {
     },
     'storage': {
         'value': 'full',
+        'name': None,
         'com': 'flag indicating which time steps to store',
         'dimension': None,
         'units': None,
@@ -151,6 +159,7 @@ _DEF_PARAM = {
     },
     'save': {
         'value': True,
+        'name': None,
         'com': 'flag indicating whether to save output data',
         'dimension': None,
         'units': None,
@@ -163,6 +172,7 @@ _DEF_PARAM = {
     # Population evolution
     'beta': {
         'value': 0.025,
+        'name': None,
         'com': 'Rate of population growth',
         'dimension': 'time rate',
         'units': 'y^{-1}',
@@ -172,6 +182,7 @@ _DEF_PARAM = {
     },
     'alpha': {
         'value': 0.02,
+        'name': None,
         'com': 'Rate of productivity increase',
         'dimension': 'time rate',
         'units': 'y^{-1}',
@@ -184,6 +195,7 @@ _DEF_PARAM = {
     # Capital properties
     'delta': {
         'value': 0.005,
+        'name': None,
         'com': 'Rate of capital depletion',
         'dimension': 'time rate',
         'units': 'y^{-1}',
@@ -196,6 +208,7 @@ _DEF_PARAM = {
     # Production
     'nu': {
         'value': 3,
+        'name': None,
         'com': 'Kapital to output ratio',   # !! IN CES its 1/A !!',
         'dimension': None,
         'units': None,
@@ -205,6 +218,7 @@ _DEF_PARAM = {
     },
     'eta': {
         'value': 1000,
+        'name': None,
         'com': '1/(1+substituability)',     # CES parameter
         'dimension': None,
         'units': None,
@@ -214,6 +228,7 @@ _DEF_PARAM = {
     },
     'b': {
         'value': .5,
+        'name': None,
         'com': 'capital part of the production',    # CES parameter
         'dimension': None,
         'units': None,
@@ -223,6 +238,7 @@ _DEF_PARAM = {
     },
     'z': {
         'value': 1,
+        'name': None,
         'com': 'Markup on salary estimation by employer',
         'dimension': None,
         'units': None,
@@ -235,6 +251,7 @@ _DEF_PARAM = {
     # INTEREST / Price
     'r': {
         'value': .03,
+        'name': None,
         'com': 'Interest at the bank',
         'dimension': 'time rate',
         'units': 'y^{-1}',
@@ -244,6 +261,7 @@ _DEF_PARAM = {
     },
     'etaP': {
         'value': .192,
+        'name': None,
         'com': 'Typical rate for inflation',
         'dimension': 'time rate',
         'units': 'y^{-1}',
@@ -253,6 +271,7 @@ _DEF_PARAM = {
     },
     'muP': {
         'value': 1.3,
+        'name': None,
         'com': 'Mark-up of price',
         'dimension': None,
         'units': None,
@@ -262,6 +281,7 @@ _DEF_PARAM = {
     },
     'gammaP': {
         'value': 1,
+        'name': None,
         'com': 'Money-illusion',
         'dimension': None,
         'units': None,
@@ -274,6 +294,7 @@ _DEF_PARAM = {
     # PHILIPS CURVE (employement-salary increase)
     'phinul': {
         'value': 0.04,
+        'name': None,
         'com': 'Unemployment rate that stops salary increase (no inflation)',
         'dimension': None,
         'units': None,
@@ -286,6 +307,7 @@ _DEF_PARAM = {
     # KEEN INVESTMENT FUNCTION (profit-investment function)
     'k0': {
         'value': -0.0065,
+        'name': None,
         'com': '',
         'dimension': None,
         'units': None,
@@ -295,6 +317,7 @@ _DEF_PARAM = {
     },
     'k1': {
         'value': np.exp(-5),
+        'name': None,
         'com': '',
         'dimension': None,
         'units': None,
@@ -304,6 +327,7 @@ _DEF_PARAM = {
     },
     'k2': {
         'value': 20,
+        'name': None,
         'com': '',
         'dimension': None,
         'units': None,
@@ -316,6 +340,7 @@ _DEF_PARAM = {
     # LINEAR DIVIDENT PROFITS
     'div0': {
         'value': 0.138,
+        'name': None,
         'com': 'Part of GDP as dividends when pi=0',
         'dimension': None,
         'units': None,
@@ -325,6 +350,7 @@ _DEF_PARAM = {
     },
     'div1': {
         'value': 0.473,
+        'name': None,
         'com': 'Slope',
         'dimension': None,
         'units': None,
@@ -337,6 +363,7 @@ _DEF_PARAM = {
     # Coupling Effets (EDP)
     'g1': {
         'value': .0,
+        'name': None,
         'com': 'GLOBAL EFFECTS OF LAMBDA (Mean field)',
         'dimension': None,
         'units': None,
@@ -346,6 +373,7 @@ _DEF_PARAM = {
     },
     'g2': {
         'value': .00,
+        'name': None,
         'com': 'WITH NEIGHBORS EFFECTS OF LAMBDA (field)',
         'dimension': None,
         'units': None,
@@ -355,6 +383,7 @@ _DEF_PARAM = {
     },
     'muI': {
         'value': 0.,
+        'name': None,
         'com': '',
         'dimension': None,
         'units': "NOTDONEYET",
@@ -364,6 +393,7 @@ _DEF_PARAM = {
     },
     'muN': {
         'value': 0.,
+        'name': None,
         'com': '',
         'dimension': None,
         'units': "NOTDONEYET",
@@ -376,6 +406,7 @@ _DEF_PARAM = {
     # RELAXATION-BUFFER DYNAMICS
     'tauR': {
         'value': 2.0,
+        'name': None,
         'com': 'Typical time for recruitement',
         'dimension': 'time',
         'units': 'y',
@@ -385,6 +416,7 @@ _DEF_PARAM = {
     },
     'tauF': {
         'value': 0.1,
+        'name': None,
         'com': 'Typical time for firing',
         'dimension': 'time',
         'units': 'y',
@@ -394,6 +426,7 @@ _DEF_PARAM = {
     },
     'tauL': {
         'value': 2.,
+        'name': None,
         'com': 'Typical time for employement information',
         'dimension': 'time',
         'units': 'y',
@@ -403,6 +436,7 @@ _DEF_PARAM = {
     },
     'tauK': {
         'value': 2.,
+        'name': None,
         'com': 'Typical time on new capital integration',
         'dimension': 'time',
         'units': 'y',
@@ -415,6 +449,7 @@ _DEF_PARAM = {
     # GEMMES PARAMETERS
     'theta': {
         'value': 2.6,
+        'name': None,
         'com': 'Convexity on abattement cost function',
         'dimension': None,
         'units': None,
@@ -424,6 +459,7 @@ _DEF_PARAM = {
     },
     'dsigma': {
         'value': -0.001,
+        'name': None,
         'com': 'Variation rate of the growth of emission intensity',
         'dimension': 'time rate',
         'units': 'y^{-1}',
@@ -433,6 +469,7 @@ _DEF_PARAM = {
     },
     'dPBS': {
         'value': -0.005,
+        'name': None,
         'com': 'Growth rate of back-stop technology price',
         'dimension': 'time rate',
         'units': 'y^{-1}',
@@ -442,6 +479,7 @@ _DEF_PARAM = {
     },
     'dEland': {
         'value': -0.022,
+        'name': None,
         'com': 'Growth rate of land use change in CO2 emission',
         'dimension': 'time rate',
         'units': 'y^{-1}',
@@ -455,6 +493,7 @@ _DEF_PARAM = {
     # D = 1 - (1 + p['pi1']*T + p['pi2']*T**2 + p['pi3']*T**p['zeta'] )**(-1)
     'pi1': {
         'value': 0.,
+        'name': None,
         'com': 'Linear temperature impact',
         'dimension': 'temperature rate',
         'units': 'T^{-1}',
@@ -464,6 +503,7 @@ _DEF_PARAM = {
     },
     'pi2': {
         'value': .00236,
+        'name': None,
         'com': 'Quadratic temperature impact',
         'dimension': None,
         'units': 'T^{-2}',
@@ -473,6 +513,7 @@ _DEF_PARAM = {
     },
     'pi3': {
         'value': .00000507,
+        'name': None,
         'com': 'Weitzmann Damage temperature impact',
         'dimension': None,
         'units': 'T^{-zeta}',
@@ -482,6 +523,7 @@ _DEF_PARAM = {
     },
     'zeta': {
         'value': 6.754,
+        'name': None,
         'com': 'Weitzmann impact',
         'dimension': None,
         'units': None,
@@ -491,6 +533,7 @@ _DEF_PARAM = {
     },
     'fk': {
         'value': 1. / 3.,
+        'name': None,
         'com': 'Fraction of environmental damage',
         'dimension': None,
         'units': None,
@@ -505,6 +548,7 @@ _DEF_PARAM = {
     # Climate model
     'Phi12': {
         'value': .024,
+        'name': None,
         'com': 'Transfer of carbon from atmosphere to biosphere',
         'dimension': None,
         'units': None,
@@ -514,6 +558,7 @@ _DEF_PARAM = {
     },
     'Phi23': {
         'value': .001,
+        'name': None,
         'com': 'Transfer from biosphere to stock',
         'dimension': None,
         'units': None,
@@ -523,6 +568,7 @@ _DEF_PARAM = {
     },
     'C': {
         'value': 1 / .098,
+        'name': None,
         'com': 'Heat capacity of fast-paced climate',
         'dimension': None,
         'units': "SI",
@@ -532,6 +578,7 @@ _DEF_PARAM = {
     },
     'C0': {
         'value': 3.52,
+        'name': None,
         'com': 'Heat capacity of inertial component of climate',
         'dimension': None,
         'units': "SI",
@@ -541,6 +588,7 @@ _DEF_PARAM = {
     },
     'gammaHEAT': {
         'value': 0.0176,
+        'name': None,
         'com': 'Heat exchange coefficient between layer',
         'dimension': None,
         'units': None,
@@ -550,6 +598,7 @@ _DEF_PARAM = {
     },
     'Tsens': {
         'value': 3.1,
+        'name': None,
         'com': 'Climate sensitivity (deltaT/log2CO2)',
         'dimension': None,
         'units': 'T',
@@ -560,6 +609,7 @@ _DEF_PARAM = {
 
     'FexoMax': {
         'value': 0.7,
+        'name': None,
         'com': 'Maximal exougenous radiative forcing',
         'dimension': None,
         'units': 'W M^{-2}',
@@ -569,6 +619,7 @@ _DEF_PARAM = {
     },
     'F2CO2': {
         'value': 3.681,
+        'name': None,
         'com': 'doubling CO2 impact on forced radiations',
         'dimension': None,
         'units': 'W/m2',
@@ -579,6 +630,7 @@ _DEF_PARAM = {
 
     'PopSat': {
         'value': 12,
+        'name': None,
         'com': 'Maximal population (billions)',
         'dimension': None,
         'units': 'Humans',
@@ -595,10 +647,9 @@ _DEF_PARAM = {
 #       _DEF_PARAM: Fill in default values and check conformity
 # #############################################################################
 
-
 _dfail = {}
 _lkeys = [
-    'value', 'com', 'dimension', 'units', 'type', 'symbol', 'group',
+    'value', 'name', 'com', 'dimension', 'units', 'type', 'symbol', 'group',
 ]
 for k0, v0 in _DEF_PARAM.items():
 
@@ -607,6 +658,12 @@ for k0, v0 in _DEF_PARAM.items():
     if len(lout) > 0:
         _dfail[k0] = f"missing keys: {lout}"
         continue
+
+    # If com if filled but not name, use com to fill name (and vice-versa)
+    if v0['name'] is None and v0['com'] is not None:
+        _DEF_PARAM[k0]['name'] = v0['com']
+    elif v0['name'] is not None and v0['com'] is None:
+        _DEF_PARAM[k0]['com'] = v0['name']
 
     # Try to spot any typo / mistake
     if v0['dimension'] not in _DALLOWED['dimension'] + [None]:
@@ -631,12 +688,16 @@ if len(_dfail) > 0:
 
 
 _DPARAM = {
-    'GK': {k0: dict(v0) for k0, v0 in _DEF_PARAM.items()},
-    'GK-reduced': {k0: dict(v0) for k0, v0 in _DEF_PARAM.items()},
+    'v0': {k0: dict(v0) for k0, v0 in _DEF_PARAM.items()},
+    'v1': {k0: dict(v0) for k0, v0 in _DEF_PARAM.items()},
+    'GreatAuthor2019': {k0: dict(v0) for k0, v0 in _DEF_PARAM.items()},
 }
 
 
-# Modify if necessary
+# Modify
+v0 = 'GreatAuthor2019'
+_DPARAM[v0]['b'] = 0.
+_DPARAM[v0]['eta'] = 0.192
 
 
 # #############################################################################
